:truck: [pkg/test] Move (Assert)Terminates(Ctx) to [pkg/context(/test)]
